roundtripper files, added tests
